+++ hutool-core/src/main/java/cn/hutool/core/exceptions/CheckedUtil.java
@@ -54,7 +54,7 @@ public class CheckedUtil {
 	 * @return {@link FuncRt}
 	 */
 	public static <P, R> FuncRt<P, R> uncheck(Func<P, R> expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
 	/**
@@ -66,7 +66,7 @@ public class CheckedUtil {
 	 * @return {@link Func0Rt}
 	 */
 	public static <R> Func0Rt<R> uncheck(Func0<R> expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
 	/**
@@ -79,7 +79,7 @@ public class CheckedUtil {
 	 * @return {@link Func1Rt}
 	 */
 	public static <P, R> Func1Rt<P, R> uncheck(Func1<P, R> expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
 
@@ -92,7 +92,7 @@ public class CheckedUtil {
 	 * @return {@link VoidFuncRt}
 	 */
 	public static <P> VoidFuncRt<P> uncheck(VoidFunc<P> expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
 	/**
@@ -103,7 +103,7 @@ public class CheckedUtil {
 	 * @return {@link VoidFunc0Rt}
 	 */
 	public static VoidFunc0Rt uncheck(VoidFunc0 expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
 	/**
@@ -115,35 +115,9 @@ public class CheckedUtil {
 	 * @return {@link VoidFunc1Rt}
 	 */
 	public static <P> VoidFunc1Rt<P> uncheck(VoidFunc1<P> expression) {
-		return uncheck(expression, new RuntimeException());
+		return uncheck(expression, RuntimeException::new);
 	}
 
-
-	/**
-	 * 接收一个可以转化成 cn.hutool.core.lang.func.Func的Lambda表达式，和一个RuntimeException，当执行表达式抛出任何异常的时候，都会转化成运行时异常
-	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
-	 *
-	 * @param expression Lambda表达式
-	 * @param rte        期望抛出的运行时异常
-	 * @param <P>        运行时传入的参数类型
-	 * @param <R>        最终返回的数据类型
-	 * @return {@link FuncRt}
-	 */
-	public static <P, R> FuncRt<P, R> uncheck(Func<P, R> expression, RuntimeException rte) {
-		Objects.requireNonNull(expression, "expression can not be null");
-		return t -> {
-			try {
-				return expression.call(t);
-			} catch (Exception e) {
-				if (rte == null) {
-					throw new RuntimeException(e);
-				} else {
-					rte.initCause(e);
-					throw rte;
-				}
-			}
-		};
-	}
 
 	/**
 	 * 接收一个可以转化成 cn.hutool.core.lang.func.Func的Lambda表达式，和一个可以把Exception转化成RuntimeExceptionde的表达式，当执行表达式抛出任何异常的时候，都会转化成运行时异常
@@ -165,31 +139,6 @@ public class CheckedUtil {
 					throw new RuntimeException(e);
 				} else {
 					throw rteSupplier.get(e);
-				}
-			}
-		};
-	}
-
-	/**
-	 * 接收一个可以转化成 cn.hutool.core.lang.func.Func0的Lambda表达式，和一个RuntimeException，当执行表达式抛出任何异常的时候，都会转化成运行时异常
-	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
-	 *
-	 * @param expression Lambda表达式
-	 * @param rte        期望抛出的运行时异常
-	 * @param <R>        最终返回的数据类型
-	 * @return {@link Func0Rt}
-	 */
-	public static <R> Func0Rt<R> uncheck(Func0<R> expression, RuntimeException rte) {
-		Objects.requireNonNull(expression, "expression can not be null");
-		return () -> {
-			try {
-				return expression.call();
-			} catch (Exception e) {
-				if (rte == null) {
-					throw new RuntimeException(e);
-				} else {
-					rte.initCause(e);
-					throw rte;
 				}
 			}
 		};
@@ -220,32 +169,6 @@ public class CheckedUtil {
 	}
 
 	/**
-	 * 接收一个可以转化成 cn.hutool.core.lang.func.Func1的Lambda表达式，和一个RuntimeException，当执行表达式抛出任何异常的时候，都会转化成运行时异常
-	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
-	 *
-	 * @param expression Lambda表达式
-	 * @param rte        期望抛出的运行时异常
-	 * @param <P>        运行时传入的参数类型
-	 * @param <R>        最终返回的数据类型
-	 * @return {@link Func1Rt}
-	 */
-	public static <P, R> Func1Rt<P, R> uncheck(Func1<P, R> expression, RuntimeException rte) {
-		Objects.requireNonNull(expression, "expression can not be null");
-		return t -> {
-			try {
-				return expression.call(t);
-			} catch (Exception e) {
-				if (rte == null) {
-					throw new RuntimeException(e);
-				} else {
-					rte.initCause(e);
-					throw rte;
-				}
-			}
-		};
-	}
-
-	/**
 	 * 接收一个可以转化成 cn.hutool.core.lang.func.Func1的Lambda表达式，和一个可以把Exception转化成RuntimeExceptionde的表达式，当执行表达式抛出任何异常的时候，都会转化成运行时异常
 	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
 	 *
@@ -265,31 +188,6 @@ public class CheckedUtil {
 					throw new RuntimeException(e);
 				} else {
 					throw rteSupplier.get(e);
-				}
-			}
-		};
-	}
-
-	/**
-	 * 接收一个可以转化成 cn.hutool.core.lang.func.VoidFunc的Lambda表达式，和一个可以把Exception转化成RuntimeExceptionde的表达式，当执行表达式抛出任何异常的时候，都会转化成运行时异常
-	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
-	 *
-	 * @param expression Lambda表达式
-	 * @param rte        期望抛出的运行时异常
-	 * @param <P>        运行时传入的参数类型
-	 * @return {@link VoidFuncRt}
-	 */
-	public static <P> VoidFuncRt<P> uncheck(VoidFunc<P> expression, RuntimeException rte) {
-		Objects.requireNonNull(expression, "expression can not be null");
-		return t -> {
-			try {
-				expression.call(t);
-			} catch (Exception e) {
-				if (rte == null) {
-					throw new RuntimeException(e);
-				} else {
-					rte.initCause(e);
-					throw rte;
 				}
 			}
 		};
@@ -367,32 +265,6 @@ public class CheckedUtil {
 		};
 	}
 
-
-	/**
-	 * 接收一个可以转化成 cn.hutool.core.lang.func.VoidFunc1的Lambda表达式，和一个RuntimeException，当执行表达式抛出任何异常的时候，都会转化成运行时异常
-	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
-	 *
-	 * @param expression Lambda表达式
-	 * @param rte        期望抛出的运行时异常
-	 * @param <P>        运行时传入的参数类型
-	 * @return {@link VoidFunc1Rt}
-	 */
-	public static <P> VoidFunc1Rt<P> uncheck(VoidFunc1<P> expression, RuntimeException rte) {
-		Objects.requireNonNull(expression, "expression can not be null");
-		return t -> {
-			try {
-				expression.call(t);
-			} catch (Exception e) {
-				if (rte == null) {
-					throw new RuntimeException(e);
-				} else {
-					rte.initCause(e);
-					throw rte;
-				}
-			}
-		};
-	}
-
 	/**
 	 * 接收一个可以转化成 cn.hutool.core.lang.func.VoidFunc1的Lambda表达式，和一个RuntimeException，当执行表达式抛出任何异常的时候，都会转化成运行时异常
 	 * 如此一来，代码中就不用显示的try-catch转化成运行时异常
@@ -418,6 +290,7 @@ public class CheckedUtil {
 	}
 
 	public interface FuncRt<P, R> extends Func<P, R> {
+		@SuppressWarnings("unchecked")
 		@Override
 		R call(P... parameters) throws RuntimeException;
 	}
@@ -433,6 +306,7 @@ public class CheckedUtil {
 	}
 
 	public interface VoidFuncRt<P> extends VoidFunc<P> {
+		@SuppressWarnings("unchecked")
 		@Override
 		void call(P... parameters) throws RuntimeException;
 	}