+++ hutool-core/src/test/java/cn/hutool/core/exceptions/CheckedUtilTest.java
@@ -38,7 +38,6 @@ public class CheckedUtilTest {
 		}
 	}
 
-	@SuppressWarnings("ConstantConditions")
 	@Test
 	public void functionTest() {
 		Func1<String, String> afunc = (funcParam) -> {